refactor(request): extract approval decision helpers
